LayoutTests/fast/dom/Range/getClientRects-expected.txt

Test 1
PASS rects.length is 5
PASS rects[0].left is 8
PASS rects[0].top is 8
PASS rects[0].width is 400
PASS rects[0].height is 160
PASS rects[1].left is 8
PASS rects[1].top is 19
PASS rects[1].width is 396
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 59
PASS rects[2].width is 398
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 99
PASS rects[3].width is 360
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 139
PASS rects[4].width is 306
PASS rects[4].height is 18
Test 2
PASS rects.length is 4
PASS rects[0].left is 8
PASS rects[0].top is 215
PASS rects[0].width is 396
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 255
PASS rects[1].width is 398
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 295
PASS rects[2].width is 360
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 335
PASS rects[3].width is 306
PASS rects[3].height is 18
Test 3
PASS rects.length is 4
PASS rects[0].left is 80
PASS rects[0].top is 411
PASS rects[0].width is 324
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 451
PASS rects[1].width is 398
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 491
PASS rects[2].width is 360
PASS rects[2].height is 18
PASS rects[3].left is 8
PASS rects[3].top is 531
PASS rects[3].width is 88
PASS rects[3].height is 18
Test 4
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 596
PASS rects[0].width is 400
PASS rects[0].height is 278
PASS rects[1].left is 8
PASS rects[1].top is 607
PASS rects[1].width is 396
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 725
PASS rects[2].width is 242
PASS rects[2].height is 18
PASS rects[3].left is 326
PASS rects[3].top is 725
PASS rects[3].width is 44
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 765
PASS rects[4].width is 341
PASS rects[4].height is 18
PASS rects[5].left is 8
PASS rects[5].top is 805
PASS rects[5].width is 366
PASS rects[5].height is 18
PASS rects[6].left is 8
PASS rects[6].top is 845
PASS rects[6].width is 67
PASS rects[6].height is 18
Test 5
PASS rects.length is 5
PASS rects[0].left is 80
PASS rects[0].top is 921
PASS rects[0].width is 324
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 1039
PASS rects[1].width is 242
PASS rects[1].height is 18
PASS rects[2].left is 250
PASS rects[2].top is 950
PASS rects[2].width is 76
PASS rects[2].height is 103
PASS rects[3].left is 326
PASS rects[3].top is 1039
PASS rects[3].width is 44
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 1079
PASS rects[4].width is 284
PASS rects[4].height is 18
Test 6
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 1224
PASS rects[0].width is 400
PASS rects[0].height is 200
PASS rects[1].left is 8
PASS rects[1].top is 1235
PASS rects[1].width is 396
PASS rects[1].height is 18
PASS rects[2].left is 8
PASS rects[2].top is 1275
PASS rects[2].width is 57
PASS rects[2].height is 18
PASS rects[3].left is 105
PASS rects[3].top is 1275
PASS rects[3].width is 44
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 1315
PASS rects[4].width is 387
PASS rects[4].height is 18
PASS rects[5].left is 8
PASS rects[5].top is 1355
PASS rects[5].width is 397
PASS rects[5].height is 18
PASS rects[6].left is 8
PASS rects[6].top is 1395
PASS rects[6].width is 171
PASS rects[6].height is 18
Test 7
PASS rects.length is 7
PASS rects[0].left is 8
PASS rects[0].top is 1471
PASS rects[0].width is 396
PASS rects[0].height is 18
PASS rects[1].left is 8
PASS rects[1].top is 1500
PASS rects[1].width is 400
PASS rects[1].height is 40
PASS rects[2].left is 8
PASS rects[2].top is 1511
PASS rects[2].width is 57
PASS rects[2].height is 18
PASS rects[3].left is 105
PASS rects[3].top is 1511
PASS rects[3].width is 44
PASS rects[3].height is 18
PASS rects[4].left is 8
PASS rects[4].top is 1551
PASS rects[4].width is 387
PASS rects[4].height is 18
PASS rects[5].left is 8
PASS rects[5].top is 1591
PASS rects[5].width is 397
PASS rects[5].height is 18
PASS rects[6].left is 8
PASS rects[6].top is 1631
PASS rects[6].width is 171
PASS rects[6].height is 18
Test 8
PASS rects.length is 4
PASS rects[0].left is 168
PASS rects[0].top is 1678
PASS rects[0].width is 269
PASS rects[0].height is 316
PASS rects[1].left is 137
PASS rects[1].top is 1704
PASS rects[1].width is 271
PASS rects[1].height is 317
PASS rects[2].left is 107
PASS rects[2].top is 1729
PASS rects[2].width is 246
PASS rects[2].height is 289
PASS rects[3].left is 76
PASS rects[3].top is 1755
PASS rects[3].width is 212
PASS rects[3].height is 247
PASS successfullyParsed is true

TEST COMPLETE
